# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cashbook Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Cashbook {
  id: number;
  name: string;
  description?: string | null;
  createdAt: string;
  updatedAt: string;
  /** Current net balance (cashIn - cashOut) */
  totalBalance: number;
  totalCashIn: number;
  totalCashOut: number;
}

export interface CreateCashbookRequest {
  name: string;
  description?: string | null;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  cash_in: "cash_in",
  cash_out: "cash_out",
} as const;

export interface Transaction {
  id: number;
  cashbookId: number;
  type: TransactionType;
  amount: number;
  particular: string;
  date: string;
  createdAt: string;
}

export type TransactionWithBalanceType =
  (typeof TransactionWithBalanceType)[keyof typeof TransactionWithBalanceType];

export const TransactionWithBalanceType = {
  cash_in: "cash_in",
  cash_out: "cash_out",
} as const;

export interface TransactionWithBalance {
  id: number;
  cashbookId: number;
  type: TransactionWithBalanceType;
  amount: number;
  particular: string;
  date: string;
  createdAt: string;
  runningBalance: number;
}

export interface TransactionListResponse {
  transactions: TransactionWithBalance[];
  totalBalance: number;
  totalCashIn: number;
  totalCashOut: number;
}

export type CreateTransactionRequestType =
  (typeof CreateTransactionRequestType)[keyof typeof CreateTransactionRequestType];

export const CreateTransactionRequestType = {
  cash_in: "cash_in",
  cash_out: "cash_out",
} as const;

export interface CreateTransactionRequest {
  type: CreateTransactionRequestType;
  amount: number;
  particular: string;
  date: string;
}

export type ListTransactionsParams = {
  search?: string;
};