# lib/api-zod/src/generated/types/cashbook.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cashbook Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Cashbook {
  id: number;
  name: string;
  description?: string | null;
  createdAt: Date;
  updatedAt: Date;
  /** Current net balance (cashIn - cashOut) */
  totalBalance: number;
  totalCashIn: number;
  totalCashOut: number;
}