# lib/api-zod/src/generated/types/createCashbookRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cashbook Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCashbookRequest {
  name: string;
  description?: string | null;
}