# lib/api-zod/src/generated/types/createTransactionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cashbook Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionRequestType } from "./createTransactionRequestType";

export interface CreateTransactionRequest {
  type: CreateTransactionRequestType;
  amount: number;
  particular: string;
  date: Date;
}