# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cashbook Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from "./transactionType";

export interface Transaction {
  id: number;
  cashbookId: number;
  type: TransactionType;
  amount: number;
  particular: string;
  date: Date;
  createdAt: Date;
}