# lib/api-zod/src/generated/types/transactionListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cashbook Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionWithBalance } from "./transactionWithBalance";

export interface TransactionListResponse {
  transactions: TransactionWithBalance[];
  totalBalance: number;
  totalCashIn: number;
  totalCashOut: number;
}