# lib/api-zod/src/generated/types/transactionWithBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cashbook Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionWithBalanceType } from "./transactionWithBalanceType";

export interface TransactionWithBalance {
  id: number;
  cashbookId: number;
  type: TransactionWithBalanceType;
  amount: number;
  particular: string;
  date: Date;
  createdAt: Date;
  runningBalance: number;
}